+++ src/index.ts
@@ -482,6 +482,7 @@ Hooks.on('ready', async () => {
 
 Hooks.on('init', async () => {
     await loadTemplates([
+        'modules/pf2e-kingmaker-tools/templates/common/skills.partial.hbs',
         'modules/pf2e-kingmaker-tools/templates/camping/activity.partial.hbs',
         'modules/pf2e-kingmaker-tools/templates/camping/eating.partial.hbs',
         'modules/pf2e-kingmaker-tools/templates/kingdom/sidebar.hbs',